ML/backend/requirements.txt
Flask
flask-cors
pandas
openpyxl
Pillow
numpy
mtcnn
